Users can now make API class to create user/game, and join/create a game
